refactor: restructure 18 DTO files into 5 with nested inner classes
- Consolidate film/, films/, search/ DTOs into 5 files
- Movie (nested Modified, Created, Category, Country, Tmdb, Imdb)
- Films (nested Pagination)
- FilmDetails (nested Episode, ServerData)
- SearchResults (nested Data, Params, SeoOnPage, BreadCrumb)
- Fix CategoryResponse package to match module
- Update imports in FilmAPI and FilmMapper
