+++ core/network/src/main/java/com/drs/auralife/core/network/model/Movie.kt
@@ -1,4 +1,4 @@
-package com.drs.auralife.core.network.model.film
+package com.drs.auralife.core.network.model
 
 import com.google.gson.annotations.SerializedName
 
@@ -7,8 +7,8 @@ data class Movie(
     val modified: Modified?,
     val name: String?,
     @SerializedName("origin_name") val originName: String?,
-    @SerializedName("poster_url") var posterUrl: String?,
-    @SerializedName("thumb_url") var thumbUrl: String?,
+    @SerializedName("poster_url") val posterUrl: String?,
+    @SerializedName("thumb_url") val thumbUrl: String?,
     @SerializedName("trailer_url") val trailerUrl: String?,
     val tmdb: Tmdb?,
     val view: Int?,
@@ -20,7 +20,7 @@ data class Movie(
     val notify: String?,
     val status: String?,
     val quality: String?,
-    var content: String?,
+    val content: String?,
     val created: Created?,
     val actor: List<String>?,
     val country: List<Country>?,
@@ -33,4 +33,17 @@ data class Movie(
     @SerializedName("is_copyright") val isCopyright: Boolean?,
     @SerializedName(value = "episode_total") val episodeTotal: String?,
     @SerializedName("episode_current") val episodeCurrent: String?,
-)
+) {
+    data class Modified(val time: String?)
+    data class Created(val time: String?)
+    data class Category(val id: String, val name: String, val slug: String)
+    data class Country(val id: String, val name: String, val slug: String)
+    data class Tmdb(
+        val id: String,
+        val season: Int,
+        val type: String,
+        @SerializedName("vote_average") val voteAverage: Double,
+        @SerializedName("vote_count") val voteCount: Int,
+    )
+    data class Imdb(val id: String?)
+}
