--- core/network/src/main/java/com/drs/auralife/core/network/model/film/Category.kt
@@ -1,7 +1,0 @@
-package com.drs.auralife.core.network.model.film
-
-data class Category(
-    val id: String,
-    val name: String,
-    val slug: String,
-)
