--- core/network/src/main/java/com/drs/auralife/core/network/model/film/Country.kt
@@ -1,7 +1,0 @@
-package com.drs.auralife.core.network.model.film
-
-data class Country(
-    val id: String,
-    val name: String,
-    val slug: String,
-)
