--- core/network/src/main/java/com/drs/auralife/core/network/model/film/Created.kt
@@ -1,5 +1,0 @@
-package com.drs.auralife.core.network.model.film
-
-data class Created(
-    val time: String,
-)
