--- core/network/src/main/java/com/drs/auralife/core/network/model/film/Episode.kt
@@ -1,8 +1,0 @@
-package com.drs.auralife.core.network.model.film
-
-import com.google.gson.annotations.SerializedName
-
-data class Episode(
-    @SerializedName("server_data") val serverData: List<ServerData>,
-    @SerializedName("server_name") val serverName: String,
-)
