--- core/network/src/main/java/com/drs/auralife/core/network/model/film/FilmDetails.kt
@@ -1,8 +1,0 @@
-package com.drs.auralife.core.network.model.film
-
-data class FilmDetails(
-    val episodes: List<Episode>,
-    val movie: Movie,
-    val msg: String,
-    val status: Boolean,
-)
