--- core/network/src/main/java/com/drs/auralife/core/network/model/film/Imdb.kt
@@ -1,5 +1,0 @@
-package com.drs.auralife.core.network.model.film
-
-data class Imdb(
-    val id: Any,
-)
