--- core/network/src/main/java/com/drs/auralife/core/network/model/film/Modified.kt
@@ -1,5 +1,0 @@
-package com.drs.auralife.core.network.model.film
-
-data class Modified(
-    val time: String,
-)
