--- core/network/src/main/java/com/drs/auralife/core/network/model/film/ServerData.kt
@@ -1,11 +1,0 @@
-package com.drs.auralife.core.network.model.film
-
-import com.google.gson.annotations.SerializedName
-
-data class ServerData(
-    val filename: String,
-    @SerializedName("link_embed") val linkEmbed: String,
-    @SerializedName("link_m3u8") val linkM3u8: String,
-    val name: String,
-    val slug: String,
-)
