--- core/network/src/main/java/com/drs/auralife/core/network/model/films/Pagination.kt
@@ -1,8 +1,0 @@
-package com.drs.auralife.core.network.model.films
-
-data class Pagination(
-    val currentPage: Int,
-    val totalItems: Int,
-    val totalItemsPerPage: Int,
-    val totalPages: Int,
-)
